Апи запрос карты. Отладочная карты. Конц и Объем во float

+++ DispenserHAL_v1.0/Project/HighLvl/modules/Card.h
@@ -8,8 +8,8 @@
 
 typedef struct
 {
-    uint16_t Volume;
-    uint8_t Concentration;
+    float Volume;
+    float Concentration;
 }Task;
 
 typedef struct
